+++ test/vector/constructors/empty_constructor/vector_empty_constructor_test.dart
@@ -1,0 +1,8 @@
+import 'package:ml_linalg/dtype.dart';
+
+import 'vector_empty_constructor_test_group_factory.dart';
+
+void main() {
+  vectorEmptyConstructorTestGroupFactory(DType.float32);
+  vectorEmptyConstructorTestGroupFactory(DType.float64);
+}