+++ test/vector/methods/multiplication_operator/multiplication_operator_test.dart
@@ -1,0 +1,9 @@
+import 'package:ml_linalg/dtype.dart';
+
+import 'multiplication_operator_test_group_factory.dart';
+
+void main() {
+  vectorMultiplicationOperatorTestGroupFactory(DType.float32);
+  // uncomment it when the Matrix class supports DType.float64
+  // vectorMultiplicationOperatorTestGroupFactory(DType.float64);
+}