+++ test/vector/methods/sqrt/sqrt_test.dart
@@ -1,0 +1,8 @@
+import 'package:ml_linalg/dtype.dart';
+
+import 'sqrt_test_group_factory.dart';
+
+void main() {
+  vectorSqrtTestGroupFactory(DType.float32);
+  vectorSqrtTestGroupFactory(DType.float64);
+}