+++ test/vector/subvector/vector_subvector_test.dart
@@ -1,7 +1,0 @@
-import 'package:ml_linalg/dtype.dart';
-
-import 'vector_subvector_test_group_factory.dart';
-
-void main() {
-  vectorSubvectorTestGroupFactory(DType.float32);
-}